tests/wpt/web-platform-tests/properties-and-values/family.html

<!doctype html>
<meta charset=utf-8>
<title>Bug 1273706 - Inheritance &amp; Cacading (CSS Properties &amp; Values)</title>
<link rel="help" href="https://drafts.css-houdini.org/css-properties-values-api/" />
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<style type="text/css">
div#container {
  font-size: 16px;
}
</style>

<div id="container">
  <div id="grandparent">
    <div id="parent">
      <div id="child"></div>
    </div>
  </div>
</div>

<script>
let grandparent = document.getElementById("grandparent");
let parent = document.getElementById("parent");
let child = document.getElementById("child");

// { registrations: [
//     { name: <custom property name with '--' prefix>,
//       syntax: <custom property syntax>,
//       inherits: bool,
//       initialValue: <computationally idempotent initial value>
//     },
//     ...
//   ],
//   grandparent: <style for grandparent element>,
//   parent: <style for parent element>,
//   child: <style for child element>,
//   expected: {
//     <computed property>: <expected value string>,
//     ...
//   },
// }
let familyTests = [
  // Basic uninherited property.
  { registrations: [
      { name: "--a",
        syntax: "<color>",
        inherits: false,
        initialValue: "red",
      },
    ],
    grandparent: "--a: green;",
    expected: {
      "--a": "red",
    },
  },
  // Transform function arguments should be computed.
  { registrations: [
      { name: "--a",
        syntax: "<transform-function>",
        inherits: true,
      },
    ],
    parent: "font-size: 10px;",
    child: "--a: translateX(5em);",
    expected: {
      "--a": "translateX(50px)",
    },
  },
  // If no initialValue is provided and the syntax is *, then a special
  // initial value used. This initial value must be considered parseable by
  // registerProperty() but invalid at computed value time.
  { registrations: [
      { name: "--a",
        syntax: "*",
        inherits: true,
      },
    ],
    parent: "--b: happy birthday; --b: hi var(--a);",
    expected: {
      // 'happy birthday' is overwritten in the cascade, and the var(--a)
      // substitution is illegal.
      "--b": "",
    },
  },
  // Same as the previous test, with a fallback (i.e. we shouldn't quit before
  // resolving).
  { registrations: [
      { name: "--a",
        syntax: "*",
        inherits: true,
      },
    ],
    parent: "--b: happy birthday; --b: hi var(--a,to you);",
    expected: {
      // The leading space is part of the value!
      "--b": " hi to you",
    },
  },
  // Make sure we calculate font-relative lengths in lengths and
  // length-percentages.
  { registrations: [
      { name: "--a",
        syntax: "<length>",
      },
      { name: "--b",
        syntax: "<length-percentage>",
      },
    ],
    parent: "font-size: 15px",
    child: "--a: 1em; --b: calc(1em + 20%);",
    expected: {
      "--a": "15px",
      "--b": "calc(15px + 20%)",
    },
  },
  // Make sure we set initial values on the root, even if nothing is ever
  // specified.
  { registrations: [
      { name: "--a",
        syntax: "<number>",
        initialValue: "42",
      },
    ],
    expected: {
      "--a": "42",
    },
  },
  // We should parse 'inherit' correctly in token stream values.
  { registrations: [
      { name: "--a",
        syntax: "*",
        inherits: false,
      },
    ],
    parent: "--a: some-value",
    child: "--a: inherit",
    expected: {
      "--a": " some-value",
    },
  },
  // We should parse 'inherit' correctly in list values.
  { registrations: [
      { name: "--a",
        syntax: "<number>+",
        inherits: false,
      },
    ],
    parent: "--a: 4 8 15",
    child: "--a: inherit",
    expected: {
      "--a": "4 8 15",
    },
  },
];

for (let spec of familyTests) {
  test(function () {
    let registered = [];
    try {
      for (let registration of spec.registrations) {
        CSS.registerProperty(registration);
        registered.push(registration.name);
      }

      grandparent.setAttribute("style", spec.grandparent || "");
      parent.setAttribute("style", spec.parent || "");
      child.setAttribute("style", spec.child || "");

      for (let key in spec.expected) {
        assert_equals(window.getComputedStyle(child).getPropertyValue(key),
          spec.expected[key], "property value equals expected value");
      }
    } finally {
      grandparent.removeAttribute("style");
      parent.removeAttribute("style");
      child.removeAttribute("style");

      for (let name of registered) {
        CSS.unregisterProperty(name);
      }
    }
  }, "family: " + JSON.stringify(spec));
}
</script>
